chore: turn coversation page into a component
